Refactor Markdown event parser (#28)
The biggest change introduced by this commit is the refactored event
parsing function that is now recursive and builds the markdown node tree
as deep as it is required for instance block quotes can be nested in
Markdown. Previous implementation did that support this easily and was
not very practical to be building up from, thus the logic was changed to
something that hopefully is easier to modify in the future.

The another big change here is the move of Markdown parser and markdown
widget directly under basalt workspace. This enables faster iteration as
I don't have to release the other workspaces constantly. So many things
are still in flux, so I want to keep the mundane things minimal.
Eventually these modules will be moved under the basalt-core and
basalt-widgets crates.

The original implementation for Markdown parser and view were kept in
place for now.

There are some other things that had to be changed due to this
modification, like the `ItemKind` and `TaskListItem` variants. These
changes were reflected to the markdown widget as well.

* Fix clippy error with matches! expression

Changelog: fixed
Package: basalt-core

* Treat clippy warnings as errors in test.yml workflow and pre-push script