Add support for `LooselyChecked` task kind
The loosely checked task type references to Obsidian flavor of
checkmarking tasks with any other character than `x`.

```markdown
- [x] Properly checked
- [ ] Unchecked
- [-] Loosely checked
- [.] Loosely checked
```

The loosely checked task item is determined in the text event, since
it's not parsed by pulldown-cmark parser. The implementation itself
reads the first four characters of the text payload in the text event
and then compares it to the following character sequence `[,_,], ` (space
is important so that we trim it away, and it's actually a valid task
item). If we match we define the current node as loosely checked task
item.