Add text wrapping for paragraphs
The paragraphs are now properly wrapped and dynamic with the size of the
terminal. The wrapping also works with different prefixes like with the
block quote prefix: `┃ `. A new dependency called `textwrap` was
introduced to wrap the strings into lines.

Added one character padding to the note pane to improve readability.

Changelog: added
Package: basalt
Module: markdown