Fix code block rendering
I falsely presumed that `String::len` returns the length as characters.
This is not the case as stated in the documentation. The `String::len`
returns the length of the string in bytes, and not in chars or
graphemes.

Changed to `str.chars().count()`, which returns the correct amount of
actual characters.

Previously, the code block might not fill the whole width of the
available area, making some lines appear to have less filled area.

Changelog: fixed
Package: basalt
Module: markdown