Functional changes to config implementation
There were some modifications that had to be implemented. One of them
was the use of BTreeMap to keep the order of key bindings the same, not
only for testing, but for later when there is a need to dynamically show
key bindings as strings. The key value of the BTreeMap was changed to
String as KeyCode does not implement Ord.

This is an intermediate change to merge the PR to main branch.

Some 'simplifications' were done as I need to continue to iterate on
this to match it with the new pane <> key binding structure, which was
not yet implemented in this commit.

The panicking default implementation was removed in favor of using the
default as a safe function. The 'panic' should happen in another layer,
for example with static-toml dependency.

Changed 'build' term to 'load' instead, and added clarifying comments
about the priority order of config loading.

System overrides > User config > Base config