Change pre-push package to allow dirty (modified, but not committed)
The last step `cargo package` would fail the pre-push hook if any files
were modified, but not yet committed. This is too strict, as there are
use cases when you have modified files and want to push partially your
_committed_ changes.

Stashing before pre-push feels clunky and unergonomic thus, changing the
command to allow dirty files to exist (modified, but not committed).