Support expandable explorer commands in Explorer widget
The expandable commands are `hide_panel` and `expand_panel`. Fixes #107.

Explorer now supports three visibility modes: hidden, visible and
full-width. Full-width takes the complete panel width from the
note-editor and fills the available space with the explorer widget.

If explorer is expanded a ⟹  symbol is shown for clarity of the state.

Change the toggle command to not change the activity state of the panel
as this can now be controlled with tab or shift-tab. Otherwise it's a
bit confusing to the user when the active panel was not changed by the
user.

Changelog: added
Package: basalt