Only text is crossed over for "hard checked" tasks
Hard checked tasks are the ones marked with 'x'. Previously the complete
item line was crossed over, however, this was not really visually
pleasing and does not match the Obsidian equivalent. This change makes
it so that the checkbox is not crossed over and only the text is.

Changelog: changed
Package: basalt