Add Renovate bot for automatic crate updates
Uses the GitHub action version of renovate. The renovate-action.json
contains the renovate configuration, and is partly derived from here:
https://github.com/Turbo87/renovate-config/blob/master/rust/updateToolchain.json

Renovate workflow is configured to run every Tuesday and Thursday.

The renovate itself will also target Tuesday and Thursday, since, this
is when new PRs are expected to be created. The workflow will also run
with workflow_dispatch or when commits are pushed to main branch, in
those cases, renovate should only update the existing PRs.