Add support to render filename in the editor pane
Now filename can be passed with a method set_file_name to the editor. If
file name is set the editor will render additional heading for the name.

The name is stylized with BlackBoardBold and underlined.

Example fila name heading if the file is called "Basalt":

```
𝔹𝕒𝕤𝕒𝕝𝕥
══════
```

Changelog: added
Package: basalt