Move some renovate config vars from workflow to config json
This makes the static configuration be defined in a single place, and
then the dynamic variants are controlled with the env vars of the
workflow.

Changed the PR body template to simpler variant, which for example
removed the controls, like the rebase PR checkbox. The checkbox does not
function immediately so it's quite irrelevant.