Big! Markdown parser refactor
This commit is large on purpose as this refactoring had many
implications to rest of the code inside note_editor module.

There is a regression in this commit, which breaks the rendering for
nested list items. However, there is now a better support for this in
the future.

The next refactoring needs to happen in the rendering part of the
markdown and note_editor, which was not the concern for this commit.

The markdown ast module is a collection of basalts' supported
markdown elements, which act as the intermediate layer between
pulldown-cmark and ratatui.

The parser itself uses recursion and "Stack" based approach to create
the necessary ast nodes. Text segments and inline styles are stored
separately from the container nodes. Container nodes are nodes that can
hold additional nodes or text element, like blockquote or heading.

There is a new text implementation called RichText, which is constructed
from TextSegment structs that can hold style. This will make inline
styling super trivial in the future.

The code in the parser is simplified, and hopefully means that
further changes are going to take less time to implement.

Added some preliminary snapshot tests for the parser. However, property
testing would also be appropriate. The snapshot tests are rendered using
S-expressions to make the ast trees more readable.

Changelog: changed
Package: basalt