Fix sub item rendering
The reason for sub list item rendering not working properly was two
fold.

a) The rendering did not support the new parser.
b) The parser did not work correctly.

This commit fixes the parser functionality to properly produce list
nodes and list item nodes recursively. ItemKind is stored in a vec so we
can actually keep track of all the item kinds we have encountered. This
is needed for e.g. calculating the correct number of ordered list items,
since we can have multiple levels of them.

Removed ListKind as obsolete. The parser now handles storing the index
of the ordered list and delegates the ownership directly to the items
themselves. This makes rendering easier as the data is readily
available.

New render_item function was introduced to Editor, that will recursively
render the list item nodes.

Changelog: fixed
Package: basalt