Add Cursor module
Cursor module is responsible for keeping up with the cursor state.
Cursor can be switched between two modes, read and edit.

Each mode behaves a bit differently, read only considers the virtual
elements, and edit mode considers the source content.

For now only the read mode variant is properly implemented.

The rendering is handled with a separate stateful CursorWidget component
that takes the cursor state as an input and draws the cursor
accordingly.

In read mode the cursor is drawn as a full-width line cursor.

Changelog: added
Package: basalt