Implement custom text editor
This custom implementation replaces tui-textarea with proper text
wrapping and better WYSIWYG experience. Also the custom implementation
allows for more granular control over how elements are, rendered and
positioned.

I tried to mimic the previous functionality in a way so as little is
lost as possible feature wise, obviously, since this is made from
scratch the feature parity is far off still.

Additionally there is some known issues, like cursor positioning is
incorrect with unicode symbols in source content. This can be observed
by wrong end position when moving the cursor to the right most end. The
cursor appears as if it is stuck, but that is due to wrong count
somewhere, which should be fixed, but in a different commit.

For now the implementation is very limited and implements only a
restricted set of features: Editing markdown nodes, saving changes to
file, moving by words and scrolling by half pages.

This commit has quite many changes, and, unfortunately the nature of how
this refactor was introduced, was difficult to separate the commits
atomically and cleanly to smaller pieces.

But the most notable ones are:

- Cursor changes, which include the cursor movement by columns and words
  and proper cursor positioning from the source offset location.

- Editor state changes to allow insertion, deletion and saving of files,
  and source range shifting, which is related to the editor
  functionality, which essentially shifts the end of the source range,
  if the source ranges are not shifted properly and if the text buffer
  exceeds the range start of next node the text buffer would be
  replicated and rendered in-place of the next node.

- Render changes, which introduce a new consolidated text wrapping and
  handling for newline characters in both visual and raw rendering
  modes. The new consolidated text wrapping uses the new whitespace
  preserving text wrap function. Additionally source offset for rendered
  virtual lines were fixed. Also added unicode-width dependency for
  accurate unicode character width calculations in render.

Changelog: changed