Fix make changelog command
Previously the changelog command was taking into account other tags than
the crate we are generating the changelog for. This would appear as if
the release had fewer commits than it actually had, this was fixed by
--count-tags flag that only counts a tag pattern like 'basalt-core/v*'.
I also had to add the --topo-order to sort the tags topologically to
ensure correct generation results.