Add Renovate auto-approve workflow
The workflow will automatically approve renovate authored PRs that have
auto-merge enabled. Auto-merge is enabled by renovate if the dependency
matches the auto-merge checks. Renovate itself enables the auto-merging
for PRs.

A new app renovate-pr-approver[bot] was added as a CODEOWNER so the PRs
can actually be approved. renovate-updater bot can not approve it's own
PRs so additional app was needed, and unfortunately that was the only
viable way to enable auto-approving.