Add functions for creating folders under obsidian vault
Introduces directory creation functions to the vault API with
`create_dir` and `create_untitled_dir`. The `find_available_note_name`
was changed to work with any kind of paths and was renamed to
`find_available_path_name`.

Add new error variant `InvalidPathName` to represents error cases where
the path is invalid. This currently happens for example if the OsStr is
not valid unicode.

Changelog: breaking