Add move functions for notes and directories in vault
Implements `move_note_to` and `move_dir_to` functions for relocating vault
filesystem entries between directories.

Add TryFrom implementations for Note and Directory structs to validate
that paths correspond to their expected filesystem types when
constructing these structs and changed the visibility of the struct
fields to ensure correctness related to filesystem type. Is it a dir or
file, which is now governed by the TryFrom implementation.

The name and path fields can now be accessed by the corresponding name()
and path() accessor methods.

Changelog: breaking